♻️ [refactor] : 가드 폴더 위치 변경
