+++ nestjs-migration/src/auth/auth.module.ts
@@ -1,16 +1,11 @@
-import { Module, forwardRef } from "@nestjs/common";
+import { Module } from "@nestjs/common";
 import { ConfigModule } from "@nestjs/config";
 import { JwtModule } from "@nestjs/jwt";
-import { UserModule } from "../user/user.module";
 import { AuthService } from "./auth.service";
 import { RefreshTokensRepository } from "./refresh-tokens.repository";
 
 @Module({
-	imports: [
-		JwtModule.register({}),
-		ConfigModule,
-		forwardRef(() => UserModule),
-	],
+	imports: [JwtModule.register({}), ConfigModule],
 	providers: [AuthService, RefreshTokensRepository],
 	exports: [AuthService, RefreshTokensRepository],
 })
